+++ src/main/ng/src/app/main/content/pages/appraisal/internalRiskRating/term-loan-risk-rating-list/term-loan-risk-rating-list.component.ts
@@ -102,12 +102,14 @@ export class TermLoanRiskRatingListComponent {
         const dialogRef = this._matDialog.open(ConfirmationDialogComponent);
         // Subscribe to the dialog close event to intercept the action taken.
         dialogRef.afterClosed().subscribe((result) => {
-            this._loanAppraisalService.deleteTermLoanRiskRating(this.selectedRating.id).subscribe(() => {
-                this._loanAppraisalService.getTermLoanRiskRatings(this._loanAppraisalId).subscribe(response => {
-                    this.dataSource = new MatTableDataSource(response._embedded.termLoanRiskRatings);
-                    this.selectedRating = undefined;
+            if (result && result.response) {
+                this._loanAppraisalService.deleteTermLoanRiskRating(this.selectedRating.id).subscribe(() => {
+                    this._loanAppraisalService.getTermLoanRiskRatings(this._loanAppraisalId).subscribe(response => {
+                        this.dataSource = new MatTableDataSource(response._embedded.termLoanRiskRatings);
+                        this.selectedRating = undefined;
+                    });
                 });
-            });
+            }
         });
     }
 }
