+++ src/main/ng/src/app/main/content/pages/monitoring/lfaList/lfaList.component.ts
@@ -133,8 +133,8 @@ export class LFAListComponent {
     deleteLFA(): void {
         const dialogRef = this._matDialog.open(ConfirmationDialogComponent);
         // Subscribe to the dialog close event to intercept the action taken.
-        dialogRef.afterClosed().subscribe((response) => {
-            if (response) {
+        dialogRef.afterClosed().subscribe((result) => {
+            if (result && result.response) {
                 this._loanMonitoringService.deleteLFA(this.selectedLFA, this._module).subscribe(() => {
                     this.selectedLFA = undefined;
                     this._loanMonitoringService.getLendersFinancialAdvisors(this.loanApplicationId).subscribe(data => {
